+++ scripts/measure/remote_capture.py
@@ -32,7 +32,6 @@ import subprocess
 import cv2
 from pprint import pprint
 import matplotlib.pyplot as plt
-import rawpy
 from lensless.hardware.utils import check_username_hostname
 from lensless.hardware.sensor import SensorOptions, sensor_dict, SensorParam
 from lensless.utils.image import rgb2gray, print_image_info
@@ -127,8 +126,11 @@ def liveview(config):
         and "bullseye" in result_dict["RPi distribution"]
         and not legacy
     ):
+        assert not rgb or not gray, "RGB and gray not supported for RPi HQ sensor"
 
         if bayer:
+
+            assert config.capture.down is None
 
             # copy over DNG file
             remotefile = f"~/{remote_fn}.dng"
@@ -137,35 +139,6 @@ def liveview(config):
             os.system('scp "%s@%s:%s" %s' % (username, hostname, remotefile, localfile))
 
             img = load_image(localfile, verbose=True, bayer=bayer, nbits_out=nbits_out)
-
-            # raw = rawpy.imread(localfile)
-
-            # # https://letmaik.github.io/rawpy/api/rawpy.Params.html#rawpy.Params
-            # # https://www.libraw.org/docs/API-datastruct-eng.html
-            # if nbits_out > 8:
-            #     # only 8 or 16 bit supported by postprocess
-            #     if nbits_out != 16:
-            #         print("casting to 16 bit...")
-            #     output_bps = 16
-            # else:
-            #     if nbits_out != 8:
-            #         print("casting to 8 bit...")
-            #     output_bps = 8
-            # img = raw.postprocess(
-            #     adjust_maximum_thr=0,  # default 0.75
-            #     no_auto_scale=False,
-            #     # no_auto_scale=True,
-            #     gamma=(1, 1),
-            #     output_bps=output_bps,
-            #     bright=1,  # default 1
-            #     exp_shift=1,
-            #     no_auto_bright=True,
-            #     # use_camera_wb=True,
-            #     # use_auto_wb=False,
-            #     # -- gives better balance for PSF measurement
-            #     use_camera_wb=False,
-            #     use_auto_wb=True,  # default is False? f both use_camera_wb and use_auto_wb are True, then use_auto_wb has priority.
-            # )
 
             # print image properties
             print_image_info(img)
@@ -240,7 +213,7 @@ def liveview(config):
 
     # save image as viewable 8 bit
     fp = os.path.join(save, f"{fn}_8bit.png")
-    save_image(img, fp)
+    save_image(img, fp, normalize=True)
 
     # plot RGB
     if plot: