+++ CHANGELOG.rst
@@ -13,6 +13,26 @@ Unreleased
 Added
 ~~~~~
 
+- Option to pass background image to ``utils.io.load_data``.
+- Option to set image resolution with ``hardware.utils.display`` function.
+
+Changed
+~~~~~~~
+
+- Nothing
+
+Bugfix
+~~~~~~
+
+- Nothing
+
+
+1.0.7 - (2024-05-14)
+--------------------
+
+Added
+~~~~~
+
 - Script to upload measured datasets to Hugging Face: ``scripts/data/upload_dataset_huggingface.py``
 - Pytorch support for simulating PSFs of masks.
 - ``lensless.hardware.mask.MultiLensArray`` class for simulating multi-lens arrays.
@@ -37,6 +57,8 @@ Changed
 - For trainable masks, set trainable parameters inside the child class.
 - ``distance_sensor`` optional for ``lensless.hardware.mask.Mask``, e.g. don't need for fabrication.
 - More intuitive interface for MURA for coded aperture (``lensless.hardware.mask.CodedAperture``), i.e. directly pass prime number.
+- ``is_torch`` to ``use_torch`` in ``lensless.hardware.mask.Mask``
+- ``self.height_map`` as characterization of phase masks (instead of phase pattern which can change for each wavelength)
 
 
 Bugfix